+++ R/augment-tk_augment_differences.R
@@ -25,6 +25,17 @@
 #'
 #'
 #' @seealso
+#'
+#' Augment Operations:
+#'
+#' - [tk_augment_timeseries_signature()] - Group-wise augmentation of timestamp features
+#' - [tk_augment_holiday_signature()] - Group-wise augmentation of holiday features
+#' - [tk_augment_roll_apply()] - Group-wise augmentation of rolling functions
+#' - [tk_augment_lags()] - Group-wise augmentation of lagged data
+#' - [tk_augment_differences()] - Group-wise augmentation of differenced data
+#'
+#' Underlying Function:
+#'
 #' - [`diff_vec()`] - Underlying function that powers `tk_augment_differences()`
 #'
 #' @examples