+++ R/augment-tk_augment_holiday_signature.R
@@ -35,6 +35,16 @@
 #'
 #' @seealso
 #'
+#' Augment Operations:
+#'
+#' - [tk_augment_timeseries_signature()] - Group-wise augmentation of timestamp features
+#' - [tk_augment_holiday_signature()] - Group-wise augmentation of holiday features
+#' - [tk_augment_roll_apply()] - Group-wise augmentation of rolling functions
+#' - [tk_augment_lags()] - Group-wise augmentation of lagged data
+#' - [tk_augment_differences()] - Group-wise augmentation of differenced data
+#'
+#' Underlying Function:
+#'
 #' - [tk_get_holiday_signature()] - Underlying function that powers holiday feature generation
 #'
 #' @examples